refactor(molecules): merge SfSteps
