--- packages/vue/src/components/molecules/SfSteps/SfSteps.html
@@ -1,21 +1,0 @@
-<div class="sf-steps">
-  <div class="sf-steps__header">
-  <!-- @slot Use this slot to customise the steps -->
-    <slot name="steps" v-for="step in parsedSteps" :step-click="stepClick" v-bind="{step}">
-      <div
-        :key="step.index"
-        :class="{'sf-steps__header-step': true, 'sf-steps__header-step-done': step.done, 'sf-steps__header-step-current': step.current, 'sf-steps__header-step-disabled': step.disabled}"
-        @click="stepClick(step)"
-      >
-        <span>{{ step.step }}</span>
-      </div>
-    </slot>
-    <div
-      class="sf-steps__progress"
-      :style="{ width: progressWidth, transform: `scaleX(${progress})` }">
-    </div>
-  </div>
-  <div class="sf-steps__content">
-    <slot></slot>
-  </div>
-</div>
