--- packages/vue/src/components/molecules/SfSteps/_internal/SfStep.html
@@ -1,3 +1,0 @@
-<div class="sf-step">
-  <slot v-if="active" :index="index"></slot>
-</div>
